+++ tests/conftest.py
@@ -229,6 +229,11 @@ def solvated_neopentane_methane():
 
 
 @pytest.fixture(scope="session")
+def ghost_14_bug():
+    return sr.load_test_files("ghost_14_bug.bss")
+
+
+@pytest.fixture(scope="session")
 def zero_lj_mols():
     return sr.load_test_files("zero_lj.prm7", "zero_lj.rst7")