+++ doc/source/changelog.rst
@@ -97,6 +97,8 @@ organisation on `GitHub <https://github.com/openbiosim/sire>`__.
 
 * Update merge code to handle ``kartograf`` API changes in version 2.0.
 
+* Allocate ``ghost-14`` slot if either end-state exception scale is nonzero.
+
 `2025.4.0 <https://github.com/openbiosim/sire/compare/2025.3.0...2025.4.0>`__ - February 2026
 ---------------------------------------------------------------------------------------------